v3.6.8
erro em consumo

+++ hotel/database/migrations/2021_10_18_183016_create_tipo_quartos_table.php
@@ -16,7 +16,7 @@ class CreateTipoQuartosTable extends Migration
         Schema::create('tipo_quartos', function (Blueprint $table) {
             $table->id();
             $table->string('nome');
-            $table->integer('valor');
+            $table->float('valor');
             $table->string('tamanho');
             $table->integer('limite_pessoa');
             $table->string('descricao');
